pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "OpenEnv-compliant email triage environment for AI agent training and evaluation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.2",
    "openai>=1.51.0",
    "httpx>=0.27.2",
    "python-multipart>=0.0.12",
    "openenv-core>=0.2.0",
]

[project.scripts]
serve = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "data*", "tasks*", "server*"]
